smarter vehicle connection for simulator
