+++ .gitignore
@@ -7,6 +7,7 @@ __pycache__/
 *.py[cod]
 *$py.class
 simulator/ardupilot
+simulator/terrain
 
 # C extensions
 *.so
@@ -133,4 +134,5 @@ my_project
 *.bin
 launch.json
 settings.json
+*.store
 # End of https://www.gitignore.io/api/python
