+++ src/util.py
@@ -3,6 +3,25 @@ import json
 import datetime
 import subprocess
 import sys
+from math import cos, sin, radians
+
+def to_quaternion(roll=0.0, pitch=0.0, yaw=0.0):
+    '''
+    Convert degrees to quaternions
+    '''
+    t0_val = cos(radians(yaw * 0.5))
+    t1_val = sin(radians(yaw * 0.5))
+    t2_val = cos(radians(roll * 0.5))
+    t3_val = sin(radians(roll * 0.5))
+    t4_val = cos(radians(pitch * 0.5))
+    t5_val = sin(radians(pitch * 0.5))
+
+    w_val = t0_val * t2_val * t4_val + t1_val * t3_val * t5_val
+    x_val = t0_val * t3_val * t4_val - t1_val * t2_val * t5_val
+    y_val = t0_val * t2_val * t5_val + t1_val * t3_val * t4_val
+    z_val = t1_val * t2_val * t4_val - t0_val * t3_val * t5_val
+
+    return [w_val, x_val, y_val, z_val]
 
 def parse_configs(argv):
     """Parses the .json file given as the first command line argument.
