+++ tox.ini
@@ -10,9 +10,9 @@ recreate = False
 passenv = *
 
 [testenv:lint]
-basepython=python3.6
+basepython=python3.7
 commands = pylint --rcfile=.pylintrc examples src
 
 [testenv:vtol]
-basepython=python3.6
+basepython=python3.7
 commands = python3 -m pytest
